Manage qtree names as '/volume/qtreename' instead of just 'qtreename'
Changed property "name" so it can be /volume/qtreename" as qtree are unique at volume level not at SVM level.
The old behaviour with still work (calling with name/volume).
New behaviour will be used passing a "/volume/qtreename" or using property "volume" and property "qtname".
Changed property "volume" to default to volume name is it's specified in the "name" property.
Added a new property "qtname" that will default to the qtree name specified in the "name" property (or the "name" property itself for the hold behaviour).
Used "qtname" instead of "name" in the qtree management functions.